+++ NobatPlusDATA/Domain/Payment.cs
@@ -9,7 +9,7 @@ namespace NobatPlusDATA.Domain
 {
     public class Payment : BaseEntity
     {
-        public int BookingID { get; set; }
+        public long BookingID { get; set; }
         public long Amount { get; set; }
         public DateTime PaymentDate { get; set; }
         public string PaymentStatus { get; set; }
