+++ NobatPlusDATA/Domain/PaymentHistory.cs
@@ -9,7 +9,7 @@ namespace NobatPlusDATA.Domain
 {
     public class PaymentHistory : BaseEntity
     {
-        public int BookingID { get; set; }
+        public long BookingID { get; set; }
         public long Amount { get; set; }
         public DateTime PaymentDate { get; set; }
         public string PaymentMethod { get; set; }
